consolidated the getLatestHealth
