+++ apex/api/sockets/aggregators/getHealthStatus.js
@@ -5,12 +5,10 @@ const {
   GET_NETWORK_HEALTH,
 } = require("../rooms");
 const { emitter, ON_SOCKET_PUBLISH_EVENTS } = require("../eventBroker");
-const rp = require("request-promise");
 const models = require("../../models");
 const config = require("../../config/app.config");
 const { Prometheus } = require("../../lib/promClient");
 const utils = require("../../lib/utils");
-const health = require("../../controllers/health");
 
 let healthBody = {
   healthStatus: null,
@@ -60,74 +58,8 @@ const counter = new Prometheus.Counter({
 
 async function getHealthStatus() {
   counter.inc();
-  const healthInfoPromise = models.CurrentHealth.findOne({
-    where: {
-      processName: "HealthStat",
-    },
-    attributes: [
-      "latestHealthStatus",
-      "latestCheckTimestamp",
-      "lastFailureTimestamp",
-    ],
-  });
-  const stallInfoPromise = models.CurrentHealth.findOne({
-    where: {
-      processName: "StallStat",
-    },
-    attributes: [
-      "latestHealthStatus",
-      "latestCheckTimestamp",
-      "lastFailureTimestamp",
-    ],
-  });
 
-  const systemInfoPromise = models.CurrentHealth.findOne({
-    where: {
-      processName: "SystemInfoStat",
-    },
-    attributes: [
-      "latestHealthStatus",
-      "latestCheckTimestamp",
-      "lastFailureTimestamp",
-      "additionalInfo",
-    ],
-    raw: true,
-  });
-
-  const syncInfoPromise = models.CurrentHealth.findOne({
-    where: {
-      processName: "SyncStat",
-    },
-    attributes: [
-      "latestHealthStatus",
-      "latestCheckTimestamp",
-      "lastFailureTimestamp",
-      "additionalInfo",
-    ],
-    raw: true,
-  });
-
-  const networkHealthPromise = models.CurrentHealth.findOne({
-    where: {
-      processName: "NetworkHealthStat",
-    },
-    attributes: [
-      "latestHealthStatus",
-      "latestCheckTimestamp",
-      "lastFailureTimestamp",
-      "additionalInfo",
-    ],
-    raw: true,
-  });
-
-  const [healthInfo, stallInfo, systemInfo, syncInfo, networkHealthInfo] =
-    await Promise.all([
-      healthInfoPromise,
-      stallInfoPromise,
-      systemInfoPromise,
-      syncInfoPromise,
-      networkHealthPromise,
-    ]);
+  const [healthInfo, stallInfo, systemInfo, syncInfo, networkHealthInfo] = await utils.getLatestHealth()
 
   if (healthInfo && stallInfo && systemInfo && syncInfo) {
     healthBody = utils.consolidateHealthData(
